requirement.txt
pdfplumber
python-docx
openai
streamlit
PyPDF2
